+++ R/stampObj.R
@@ -7,10 +7,11 @@
 ##' @param data The dataset to stamp.
 ##' @param script path to the script where the dataset was generated.
 ##' @param ... other named metadata elements to add to the dataset. Example:
-##'     Description="PK data for ph1 trials in project NNXXXX".
+##'     Description="PK data for ph1 trials in project".
 ##' @seealso objInfo
 ##' @family DataWrangling
 ##' @export
+
 
 stampObj <- function(data,script,...){
     attr(data,"objInfo") <-