+++ NEWS.md
@@ -1,3 +1,4 @@
 # NEWS
 ## Post 0.4.800
+### NMwriteOutput: support for args.stamp, a list of args passed to stampObj.
 ### means: Introduced na.rm argument.