+++ NEWS.md
@@ -1,4 +1,8 @@
 # NEWS
 ## Post 0.4.800
+### findVars: New function that finds columns that vary within other columns (say subjects).
+### findCovs: Implemented in data.table. Should be much faster.
+### ggwrite: Added a show argument. write and show are independent of each other.
 ### NMwriteOutput: support for args.stamp, a list of args passed to stampObj.
 ### means: Introduced na.rm argument. 
+### signif2: bugfixes related to zero values.